+++ R/slfhelper-package.R
@@ -5,7 +5,6 @@
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start
 #' @import rlang
-#' @importFrom glue glue
 #' @importFrom lifecycle deprecated
 #' @importFrom tibble tibble
 ## usethis namespace: end