+++ tests/testthat/test-multiple_selections.R
@@ -2,12 +2,13 @@ skip_on_ci()
 
 
 test_that("select years and recid", {
+  set.seed(50)
+
   acute_only <- read_slf_episode(c("1718", "1819"),
-    columns = c("year", "anon_chi", "recid", "keydate1_dateformat"),
-    recids = "01B",
-    from = 1000,
-    to = 200000
-  )
+    col_select = c("year", "anon_chi", "recid", "keydate1_dateformat"),
+    recids = "01B"
+  ) %>%
+    dplyr::slice_sample(n = 200000)
 
   expect_equal(
     names(acute_only),
@@ -17,11 +18,10 @@ test_that("select years and recid", {
   expect_equal(unique(acute_only$recid), "01B")
 
   hosp_only <- read_slf_episode(c("1718", "1819"),
-    columns = c("year", "anon_chi", "recid", "keydate1_dateformat"),
-    recids = c("01B", "02B", "04B", "GLS"),
-    from = 1000,
-    to = 200000
-  )
+    col_select = c("year", "anon_chi", "recid", "keydate1_dateformat"),
+    recids = c("01B", "02B", "04B", "GLS")
+  ) %>%
+    dplyr::slice_sample(n = 200000)
 
   expect_equal(
     names(hosp_only),
@@ -32,74 +32,73 @@ test_that("select years and recid", {
 })
 
 test_that("select recids and partnerships", {
-  edi_acute <- read_slf_episode("1819",
-    columns = c("year", "anon_chi", "recid", "hscp2018"),
+  set.seed(50)
+
+  edi_acute <- read_slf_episode("1718",
+    col_select = c("year", "anon_chi", "recid", "hscp2018"),
     partnerships = "S37000012",
-    recid = "01B",
-    from = 1000,
-    to = 200000
-  )
+    recid = "01B"
+  ) %>%
+    dplyr::slice_sample(n = 200000)
 
   expect_equal(names(edi_acute), c("year", "anon_chi", "recid", "hscp2018"))
-  expect_equal(unique(edi_acute$year), "1819")
+  expect_equal(unique(edi_acute$year), "1718")
   expect_equal(unique(edi_acute$recid), "01B")
   expect_equal(unique(edi_acute$hscp2018), "S37000012")
 
-  edi_gla_acute <- read_slf_episode("1819",
-    columns = c("year", "anon_chi", "recid", "hscp2018"),
+  edi_gla_acute <- read_slf_episode("1718",
+    col_select = c("year", "anon_chi", "recid", "hscp2018"),
     partnerships = c("S37000012", "S37000015"),
-    recid = "01B",
-    from = 1000,
-    to = 200000
-  )
+    recid = "01B"
+  ) %>%
+    dplyr::slice_sample(n = 200000)
 
   expect_equal(
     names(edi_gla_acute),
     c("year", "anon_chi", "recid", "hscp2018")
   )
-  expect_equal(unique(edi_gla_acute$year), "1819")
+  expect_equal(unique(edi_gla_acute$year), "1718")
   expect_equal(unique(edi_gla_acute$recid), "01B")
   expect_equal(
     sort(unique(edi_gla_acute$hscp2018)),
     c("S37000012", "S37000015")
   )
 
-  edi_hosp <- read_slf_episode("1819",
-    columns = c("year", "anon_chi", "recid", "hscp2018"),
+  edi_hosp <- read_slf_episode("1718",
+    col_select = c("year", "anon_chi", "recid", "hscp2018"),
     partnerships = "S37000012",
-    recids = c("01B", "02B", "04B", "GLS"),
-    from = 1000,
-    to = 200000
-  )
+    recids = c("01B", "02B", "04B", "GLS")
+  ) %>%
+    dplyr::slice_sample(n = 200000)
 
   expect_equal(names(edi_hosp), c("year", "anon_chi", "recid", "hscp2018"))
-  expect_equal(unique(edi_hosp$year), "1819")
+  expect_equal(unique(edi_hosp$year), "1718")
   expect_equal(sort(unique(edi_hosp$recid)), c("01B", "02B", "04B", "GLS"))
   expect_equal(unique(edi_hosp$hscp2018), "S37000012")
 
-  edi_gla_hosp <- read_slf_episode("1819",
-    columns = c("year", "anon_chi", "recid", "hscp2018"),
+  edi_gla_hosp <- read_slf_episode("1718",
+    col_select = c("year", "anon_chi", "recid", "hscp2018"),
     partnerships = c("S37000012", "S37000015"),
-    recids = c("01B", "02B", "04B", "GLS"),
-    from = 1000,
-    to = 200000
-  )
+    recids = c("01B", "02B", "04B", "GLS")
+  ) %>%
+    dplyr::slice_sample(n = 200000)
 
   expect_equal(names(edi_gla_hosp), c("year", "anon_chi", "recid", "hscp2018"))
-  expect_equal(unique(edi_gla_hosp$year), "1819")
+  expect_equal(unique(edi_gla_hosp$year), "1718")
   expect_equal(sort(unique(edi_gla_hosp$recid)), c("01B", "02B", "04B", "GLS"))
   expect_equal(sort(unique(edi_gla_hosp$hscp2018)), c("S37000012", "S37000015"))
 })
 
 
 test_that("all selections", {
+  set.seed(50)
+
   edi_gla_hosp_2_year <- read_slf_episode(c("1718", "1819"),
-    columns = c("year", "anon_chi", "recid", "hscp2018"),
+    col_select = c("year", "anon_chi", "recid", "hscp2018"),
     partnerships = c("S37000012", "S37000015"),
-    recids = c("01B", "02B", "04B", "GLS"),
-    from = 1000,
-    to = 200000
-  )
+    recids = c("01B", "02B", "04B", "GLS")
+  ) %>%
+    dplyr::slice_sample(n = 200000)
 
   expect_equal(
     names(edi_gla_hosp_2_year),