+++ tests/testthat/test-recid_selection.R
@@ -2,13 +2,14 @@ skip_on_ci()
 
 
 test_that("Recid filtering works", {
+  set.seed(50)
+
   # Read in a bit of a file selecting only Edinburgh
   ep_1718_acute <- read_slf_episode("1718",
     recids = "01B",
-    columns = c("recid"),
-    from = 1000,
-    to = 100000
-  )
+    col_select = c("recid")
+  ) %>%
+    dplyr::slice_sample(n = 100000)
 
   # Should only have Edinburgh codes
   expect_true(all(ep_1718_acute$recid == "01B"))
@@ -20,13 +21,15 @@ test_that("Recid filtering works", {
 
 
 test_that("Can select multiple recids", {
+  set.seed(50)
+
   # Read in a bit of a file selecting Edinburgh and Glasgow
   ep_1718_acute <- read_slf_episode("1718",
     recids = c("01B", "02B", "04B"),
-    columns = c("recid"),
-    from = 1000,
-    to = 100000
-  )
+    col_select = c("recid")
+  ) %>%
+    dplyr::slice_sample(n = 100000)
+
   # Should only have SMRA 1, 2 and 4
   expect_equal(sort(unique(ep_1718_acute$recid)), c("01B", "02B", "04B"))
   # Should have at least 100 records (checks we're not getting an empty file)
@@ -34,14 +37,15 @@ test_that("Can select multiple recids", {
 })
 
 test_that("Can still do filtering if variable is not selected", {
+  set.seed(50)
+
   # Read in a bit of a file selecting only Edinburgh
   # Don't choose to read the partnership variable
   ep_1718_acute <- read_slf_episode("1718",
     recids = "01B",
-    columns = c("sparra_end_fy"),
-    from = 1000,
-    to = 100000
-  )
+    col_select = c("sparra_end_fy")
+  ) %>%
+    dplyr::slice_sample(n = 100000)
 
   # We shouldn't have the recid variable
   expect_false("recid" %in% names(ep_1718_acute))
@@ -54,11 +58,12 @@ test_that("Can still do filtering if variable is not selected", {
 })
 
 test_that("Still reads all variables if just filtering", {
+  set.seed(50)
+
   ep_1718_acute <- read_slf_episode("1718",
-    recids = "01B",
-    from = 1000,
-    to = 100000
-  )
+    recids = "01B"
+  ) %>%
+    dplyr::slice_sample(n = 100000)
 
   # Should only have Edinburgh codes
   expect_true(all(ep_1718_acute$recid == "01B"))