+++ NEWS.md
@@ -1,3 +1,9 @@
+# slfhelper 0.10.0
+
+* [`{glue}`](https://glue.tidyverse.org/) is no longer a dependency as the required functionality can be provided by [`stringr::str_glue()](https://stringr.tidyverse.org/reference/str_glue.html).
+* Dependency versions have been updated to the latest.
+* `get_chi()` and `get_anon_chi()` now properly match missing (`NA`) and blank (`""`) values.
+
 # slfhelper 0.9.0
 
 * [`{openssl}`](https://cran.r-project.org/web/packages/openssl/index.html) is now required.